refactor: 대형 페이지 컴포넌트 분할 + pre-commit hooks
- BookshelfPage: 384줄 → 194줄 (useBookshelf hook + BookFilters, BookGrid 분리)
- SearchPage: 372줄 → 217줄 (useBookSearch hook + SearchResults 분리)
- WriteReviewPage: useTagInput hook 추출
- ReviewDetailPage: 인라인 서브컴포넌트 분리
- ReviewForm(TagInputField) 공유 컴포넌트 생성
- .pre-commit-config.yaml 추가 (ruff, trailing-whitespace, large-file-check)
- pyproject.toml 프로젝트명 오타 수정

close #37
